+++ awake-vulkan/src/commonMain/kotlin/io/github/ronjunevaldoz/awake/vulkan/models/info/pipeline/VkPipelineVertexInputStateCreateInfo.kt
@@ -19,6 +19,7 @@
 
 package io.github.ronjunevaldoz.awake.vulkan.models.info.pipeline
 
+import io.github.ronjunevaldoz.awake.vulkan.VkArray
 import io.github.ronjunevaldoz.awake.vulkan.VkFlags
 import io.github.ronjunevaldoz.awake.vulkan.enums.VkFormat
 import io.github.ronjunevaldoz.awake.vulkan.enums.VkStructureType
@@ -41,9 +42,9 @@ class VkPipelineVertexInputStateCreateInfo(
     val sType: VkStructureType = VkStructureType.VK_STRUCTURE_TYPE_PIPELINE_VERTEX_INPUT_STATE_CREATE_INFO,
     val pNext: Any? = null,
     val flags: VkPipelineVertexInputStateCreateFlags = 0,
-//     val vertexBindingDescriptionCount: Int,
+    @VkArray(sizeAlias = "vertexBindingDescriptionCount")
     val pVertexBindingDescriptions: Array<VkVertexInputBindingDescription>? = null,
-//     val vertexAttributeDescriptionCount: Int,
+    @VkArray(sizeAlias = "vertexAttributeDescriptionCount")
     val pVertexAttributeDescriptions: Array<VkVertexInputAttributeDescription>? = null,
 )