+++ awake-vulkan/src/main/cpp/common/utils/vulkan_utils.cpp
@@ -1019,27 +1019,26 @@ namespace vulkan_utils {
         auto pipelineCache = reinterpret_cast<VkPipelineCache>(pPipelineCache);
         auto createInfoSize = env->GetArrayLength(createInfosObj);
         std::vector<VkGraphicsPipelineCreateInfo> createInfos;
-//        for (int i = 0; i < createInfoSize; ++i) {
-        auto createInfoObj = env->GetObjectArrayElement(createInfosObj, 0);
-        VkGraphicsPipelineCreateInfoAccessor accessor(env, createInfoObj);
-        VkGraphicsPipelineCreateInfo createInfo{};
-        accessor.fromObject(createInfo);
-        createInfos.push_back(createInfo);
-//        }
-//        std::vector<VkPipeline> pipelines(createInfoSize);
-        VkPipeline pipeline;
-        VkResult result = vkCreateGraphicsPipelines(device, VK_NULL_HANDLE,
-                                                    1,
-                                                    &createInfo, nullptr, &pipeline);
+        for (int i = 0; i < createInfoSize; ++i) {
+            auto createInfoObj = env->GetObjectArrayElement(createInfosObj, i);
+            VkGraphicsPipelineCreateInfoAccessor accessor(env, createInfoObj);
+            VkGraphicsPipelineCreateInfo createInfo{};
+            accessor.fromObject(createInfo);
+            createInfos.push_back(createInfo);
+        }
+        std::vector<VkPipeline> pipelines(createInfoSize);
+        VkResult result = vkCreateGraphicsPipelines(device, pipelineCache,
+                                                    static_cast<uint32_t>(createInfos.size()),
+                                                    createInfos.data(), nullptr, pipelines.data());
         if (result != VK_SUCCESS) {
             return jlongArray();
         }
-//        jlongArray pipelineHandles = env->NewLongArray((jsize) pipelines.size());
-//        for (int i = 0; i < pipelines.size(); ++i) {
-//            auto pipeline = reinterpret_cast<jlong>(pipelines[i]);
-//            env->SetLongArrayRegion(pipelineHandles, i, 1, &pipeline);
-//        }
-        return nullptr;
+        jlongArray pipelineHandles = env->NewLongArray((jsize) pipelines.size());
+        for (int i = 0; i < pipelines.size(); ++i) {
+            auto pipeline = reinterpret_cast<jlong>(pipelines[i]);
+            env->SetLongArrayRegion(pipelineHandles, i, 1, &pipeline);
+        }
+        return pipelineHandles;
     }
 
     void destroyPipeline(jlong pDevice, jlong pPipeline) {