+++ CHANGELOG.md
@@ -35,6 +35,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Known Issues
 
+- vkCreateGraphicsPipelines hangs on emulator when shader location is using dynamic and
+  pVertexInputState is null or empty
 - Vulkan validation error (VkDeviceQueueCreateInfo-pQueuePriorities ) is not between 0 and 1 (
   inclusive) actual device (Tested only in Xiaomi MI lite 12)
 - Possible memory leak when using Vulkan.